@@ -1,60 +1,45 @@
 {
-  "name": "camp-cadet-web",
+  "name": "camp-cadet",
   "version": "0.0.0",
-  "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.5",
-    "@angular/cdk": "^5.2.2",
-    "@angular/common": "^5.2.5",
-    "@angular/compiler": "^5.2.5",
-    "@angular/core": "^5.2.5",
-    "@angular/forms": "^5.2.5",
-    "@angular/http": "^5.2.5",
-    "@angular/material": "^5.2.2",
-    "@angular/material-moment-adapter": "^5.2.3",
-    "@angular/platform-browser": "^5.2.5",
-    "@angular/platform-browser-dynamic": "^5.2.7",
-    "@angular/platform-server": "^5.2.7",
-    "@angular/router": "^5.2.7",
-    "angular5-social-login": "^1.0.9",
-    "bootstrap": "^4.0.0",
-    "core-js": "^2.5.3",
+    "@angular/animations": "^17.1.0",
+    "@angular/cdk": "^17.1.2",
+    "@angular/common": "^17.1.0",
+    "@angular/compiler": "^17.1.0",
+    "@angular/core": "^17.1.0",
+    "@angular/forms": "^17.1.0",
+    "@angular/material": "^17.1.2",
+    "@angular/platform-browser": "^17.1.0",
+    "@angular/platform-browser-dynamic": "^17.1.0",
+    "@angular/router": "^17.1.0",
+    "@emailjs/browser": "^4.1.0",
     "font-awesome": "^4.7.0",
-    "hammerjs": "^2.0.8",
-    "jquery": "^3.3.1",
-    "moment": "^2.20.1",
-    "ng2-toastr": "^4.1.2",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.20"
+    "luxon": "^3.4.4",
+    "ngx-toastr": "^18.0.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.2",
-    "@angular/compiler-cli": "^5.2.7",
-    "@angular/language-service": "^5.2.5",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.96",
-    "codelyzer": "^4.1.0",
-    "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.1",
-    "karma-jasmine": "^1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "^2.6.2"
+    "@angular-devkit/build-angular": "^17.1.1",
+    "@angular/cli": "^17.1.1",
+    "@angular/compiler-cli": "^17.1.0",
+    "@types/jasmine": "~5.1.0",
+    "@types/luxon": "^3.4.2",
+    "jasmine-core": "~5.1.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "~5.3.2"
   }
 }
