@@ -1,15 +1,15 @@
-//#region Imports
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { environment } from '@environment/environment';
-import { Observable } from 'rxjs/Observable';
-import { SimpleItem } from '@campcadet/model/simple-item';
-import { Email } from '@campcadet/model/email';
-//#endregion
+import { Observable } from 'rxjs';
+import { config } from '../config';
+import { Email } from '../model/email';
+import { SimpleItem } from '../model/simple-item';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class ContactService {
-  private url = `${environment.apiUrl}/api`;
+  private url = `${config.api}/api`;
 
   constructor(private http: HttpClient) { }
 
