@@ -2,11 +2,16 @@
 import { Component } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { MatDialogRef } from '@angular/material/dialog';
+import { from } from 'rxjs';
+import emailjs from '@emailjs/browser';
 
-import { DataService } from '@campcadet/services/data.service';
-import { ContactService } from '@campcadet/services/contact.service';
-import { Email } from '@campcadet/model/email';
-import { FormErrorState } from '@campcadet/model/form-error-state';
+import { config } from '../../config';
+
+import { DataService } from '../../services/data.service';
+//import { ContactService } from '../../services/contact.service';
+//import { Email } from '../../model/email';
+import { FormErrorState } from '../../model/form-error-state';
+import { EmailJsResponseStatus } from '../../model/emailjs-response-status';
 //#endregion
 
 @Component({
@@ -31,28 +36,63 @@ export class ContactComponent {
   //#endregion
   matcher = new FormErrorState();
 
-  private sendToDirectors = false;
+  sendToDirectors = false;
   //#endregion
 
   constructor(
-    private contactService: ContactService,
+    //private contactService: ContactService,
     public dataService: DataService,
     public dialogRef: MatDialogRef<ContactComponent>) {
 
   }
 
+  sendEmailViaEmailJs(): void {
+    if (!this.dataService.configSettings) { return; }
+
+    const type = this.dataService.contactTypes
+      .find(t => t.id === +this.contactForm.value.type!)
+      ?.description;
+
+    const reason = this.dataService.contactReasons
+      .find(r => r.id === +this.contactForm.value.reason!)
+      ?.description;
+
+    const templateParams = {
+      ...this.contactForm.value,
+      type,
+      reason,
+      subject: 'Camp Cadet Website Inquiry',
+    };
+
+    const observable$ = from(emailjs.send(config.emailJs.serviceId, config.emailJs.templateId, templateParams));
+
+    observable$.subscribe({
+      next: (emailJsResponseStatus: EmailJsResponseStatus) => {
+        console.log(emailJsResponseStatus);
+        this.dialogRef.close(true);
+      },
+      error: (err) => {
+        console.log(err);
+        this.dialogRef.close(false);
+      }
+    });
+  }
+
+  /*
   sendEmail(): void {
     // 'relay-hosting.secureserver.net', // 'smtp.gmail.com',
+    if (!this.dataService.configSettings) { return; }
+
     const email: Email = {
-      server: this.dataService.configSettings.find(d => d.description.toLowerCase() === 'emailservergodaddy').value,
-      smtpPort: +this.dataService.configSettings.find(d => d.description.toLowerCase() === 'emailsmtpport').value,
-      useSsl: this.dataService.configSettings.find(d => d.description.toLowerCase() === 'emailusessl').value === 'true' ?
+      server: this.dataService.configSettings.find(d => d.description.toLowerCase() === 'emailservergodaddy')!.value,
+      smtpPort: +this.dataService.configSettings.find(d => d.description.toLowerCase() === 'emailsmtpport')!.value,
+      useSsl: this.dataService.configSettings.find(d => d.description.toLowerCase() === 'emailusessl')!.value === 'true' ?
         true : false,
-      priority: +this.dataService.configSettings.find(d => d.description.toLowerCase() === 'emailpriority').value,
-      subject: this.dataService.configSettings.find(d => d.description.toLowerCase() === 'contactformsubject').value,
+      priority: +this.dataService.configSettings.find(d => d.description.toLowerCase() === 'emailpriority')!.value,
+      subject: this.dataService.configSettings.find(d => d.description.toLowerCase() === 'contactformsubject')!.value,
       sender: {
-        displayName: this.dataService.configSettings.find(d => d.description.toLowerCase() === 'emailfromname').value,
-        emailAddress: this.dataService.configSettings.find(d => d.description.toLowerCase() === 'emailfrom').value
+        displayName: this.dataService.configSettings.find(d => d.description.toLowerCase() === 'emailfromname')!.value,
+        emailAddress: this.dataService.configSettings.find(d => d.description.toLowerCase() === 'emailfrom')!.value
       },
       recipients: { to: [] },
       content: {
@@ -60,32 +100,36 @@ export class ContactComponent {
       }
     };
 
-    const forcedRecipients = this.dataService.configSettings.find(d => d.description.toLowerCase() === 'contactformrecipients')
-      .value.split(';');
+    const forcedRecipients = this.dataService.configSettings.find(d => d.description.toLowerCase() === 'contactformrecipients')!.value.split(';');
 
     // Send email to all board directors?
     if (this.sendToDirectors) {
       this.dataService.boardDirectors
         .filter(director => director.email)
         .map(director => director.email)
-        .forEach(address => email.recipients.to.push(address));
+        .forEach(address => email.recipients.to!.push(address!));
     }
 
     // Send email to any forced email addresses
     if (forcedRecipients && forcedRecipients.length) {
       forcedRecipients.forEach(address => {
-        if (email.recipients.to.indexOf(address) === -1) {
-          email.recipients.to.push(address);
+        if (email.recipients.to!.indexOf(address) === -1) {
+          email.recipients.to!.push(address);
         }
       });
     }
 
     this.contactService.sendEmail(email)
-      .subscribe(() => this.dialogRef.close(true),
-        err => {
+      .subscribe({
+        next: (response) => {
+          console.log(response);
+          this.dialogRef.close(true);
+        },
+        error: (err) => {
           console.log(err);
           this.dialogRef.close(false);
-        });
+        }
+      });
   }
 
   private getStyle(): string {
@@ -95,8 +139,8 @@ export class ContactComponent {
   }
 
   private getFormData(): string {
-    const type = this.dataService.contactTypes.find(t => t.id === this.contactForm.controls['type'].value);
-    const reason = this.dataService.contactReasons.find(r => r.id === this.contactForm.controls['reason'].value);
+    const type = this.dataService.contactTypes.find(t => t.id === +this.contactForm.controls['type'].value!);
+    const reason = this.dataService.contactReasons.find(r => r.id === +this.contactForm.controls['reason'].value!);
     const phone = this.contactForm.controls['phone'].value ?
       `<div class="p"><div class="bold">Phone:</div><div>${this.contactForm.controls['phone'].value}</div></div>` :
       '';
@@ -105,10 +149,9 @@ export class ContactComponent {
       <div class="p"><div class="bold">Sender:</div><div>${this.contactForm.controls['name'].value}</div></div>
       <div class="p"><div class="bold">Email:</div><div>${this.contactForm.controls['email'].value}</div></div>
       ${phone}
-      <div class="p"><div class="bold">Contact Type:</div><div>${type.description}</div></div>
-      <div class="p"><div class="bold">Contact Reason:</div><div>${reason.description}</div></div>
-      <div class="p"><div class="bold">Comment/Question:</div><div>${this.contactForm.controls['comments'].value
-        .replace(new RegExp('\n', 'g'), '<br />')}</div></div>
+      <div class="p"><div class="bold">Contact Type:</div><div>${type!.description}</div></div>
+      <div class="p"><div class="bold">Contact Reason:</div><div>${reason!.description}</div></div>
+      <div class="p"><div class="bold">Comment/Question:</div><div>${this.contactForm.controls['comments'].value!.replace(new RegExp('\n', 'g'), '<br />')}</div></div>
     `;
-  }
+  }*/
 }
