@@ -1,32 +1,26 @@
-//#region Imports
 import { Component } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
-import { AuthenticationService } from '@campcadet/services/authentication.service';
-import { BlurbEditable } from '@campcadet/model/blurb-editable';
-import { DataService } from '@campcadet/services/data.service';
-//#endregion
+import { DataService } from '../../services/data.service';
 
 @Component({
   selector: 'cc-donate',
   templateUrl: './donate.component.html',
   styleUrls: ['./donate.component.css']
 })
-export class DonateComponent extends BlurbEditable {
+export class DonateComponent {
   title = 'Donation Information';
   blurbName = 'Donations';
-  donation: number;
+  donation: number | undefined;
 
-  constructor(public dataService: DataService,
-    dialog: MatDialog,
-    public authenticationService: AuthenticationService) {
-    super(dialog, authenticationService);
-  }
+  constructor(public dataService: DataService) {}
 
   paypalSubmit(form: any, e: any): void {
     e.target.submit();
   }
 
-  calculateDonation(val): void {
-    this.donation = val ? +val + ((+val * .022) + .3) : null;
+  /*calculateDonation(val: number): void {
+    this.donation = val ? +val + ((+val * .022) + .3) : undefined;
+  }*/
+  calculateDonation(val: any): void {
+    console.log(val);
   }
 }
