@@ -1,23 +1,15 @@
-//#region Imports
 import { Component } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
-import { AuthenticationService } from '@campcadet/services/authentication.service';
-import { BlurbEditable } from '@campcadet/model/blurb-editable';
-import { DataService } from '@campcadet/services/data.service';
-//#endregion
+import { DataService } from '../../services/data.service';
 
 @Component({
   selector: 'cc-donors',
   templateUrl: './donors.component.html',
   styleUrls: ['./donors.component.css']
 })
-export class DonorsComponent extends BlurbEditable {
+export class DonorsComponent {
   blurbName = 'Donor Introduction';
   blurbNameBottom = 'Donors Bottom';
 
-  constructor(public dataService: DataService,
-    dialog: MatDialog,
-    public authenticationService: AuthenticationService) {
-    super(dialog, authenticationService);
+  constructor(public dataService: DataService) {
   }
 }
