@@ -1,23 +1,15 @@
-//#region Imports
 import { Component } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
-import { AuthenticationService } from '@campcadet/services/authentication.service';
-import { BlurbEditable } from '@campcadet/model/blurb-editable';
-import { DataService } from '@campcadet/services/data.service';
-//#endregion
+import { DataService } from '../../services/data.service';
 
 @Component({
   selector: 'cc-eligibility',
   templateUrl: './eligibility.component.html',
   styleUrls: ['./eligibility.component.css']
 })
-export class EligibilityComponent extends BlurbEditable {
+export class EligibilityComponent {
   title = 'Eligibility Requirements';
   blurbName = 'Eligibility';
 
-  constructor(public dataService: DataService,
-    dialog: MatDialog,
-    public authenticationService: AuthenticationService) {
-    super(dialog, authenticationService);
+  constructor(public dataService: DataService) {
   }
 }
