@@ -1,23 +1,14 @@
-//#region Imports
 import { Component } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
-import { AuthenticationService } from '@campcadet/services/authentication.service';
-import { BlurbEditable } from '@campcadet/model/blurb-editable';
-import { DataService } from '@campcadet/services/data.service';
-//#endregion
+import { DataService } from '../../services/data.service';
 
 @Component({
   selector: 'cc-enrollment',
   templateUrl: './enrollment.component.html',
   styleUrls: ['./enrollment.component.css']
 })
-export class EnrollmentComponent extends BlurbEditable {
+export class EnrollmentComponent {
   title = 'Camp Enrollment';
   blurbName = 'Sign Up';
 
-  constructor(public dataService: DataService,
-    dialog: MatDialog,
-    public authenticationService: AuthenticationService) {
-    super(dialog, authenticationService);
-  }
+  constructor(public dataService: DataService) {}
 }
