@@ -1,23 +1,14 @@
-//#region Imports
 import { Component } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
-import { AuthenticationService } from '@campcadet/services/authentication.service';
-import { BlurbEditable } from '@campcadet/model/blurb-editable';
-import { DataService } from '@campcadet/services/data.service';
-//#endregion
+import { DataService } from '../../services/data.service';
 
 @Component({
   selector: 'cc-preparation',
   templateUrl: './preparation.component.html',
   styleUrls: ['./preparation.component.css']
 })
-export class PreparationComponent extends BlurbEditable {
+export class PreparationComponent {
   title = 'Preparing for Camp Cadet';
   blurbName = 'Camp Preparation';
 
-  constructor(public dataService: DataService,
-    dialog: MatDialog,
-    public authenticationService: AuthenticationService) {
-      super(dialog, authenticationService);
-  }
+  constructor(public dataService: DataService) {}
 }
