@@ -1,4 +1,4 @@
-export class Blurb {
+export interface Blurb {
     id: number;
     name: string;
     blurb: string;
