@@ -1,4 +1,4 @@
-import { SimpleItem } from '@campcadet/model/simple-item';
+import { SimpleItem } from './simple-item';
 
 export interface DonorLevel extends SimpleItem {
     amountLower: number;
